See: bug 95959 - replace two different messages with a single one.
It is clearer if you state which settings need to be enabled for
a feature to work rather than have 2 messages which state different
reasons why it might fail.  This change addresses that, but should
make no functionally different behaviour apart from the change in
the error message.

Make the old error message obsolete and add a new one which is
clearer.
Adjust the test suite to look for the new message instead.